refactor: Moved some entities to public and common folders. Added separate entity for DetailedProgram
