+++ src/services/api/public/programs/programs-api.ts
@@ -1,5 +1,5 @@
 import { axiosInstance } from '@/services/api/axios';
-import { ProgramsPageData, PublishedProgramDto } from '@/types/public/programs-page';
+import { DetailedProgram, ProgramsPageData, PublishedProgramDto } from '@/types/public/programs-page';
 import { Program } from '@/types/admin/programs';
 import { API_ROUTES } from '@/const/common/api-routes/main-api';
 
@@ -16,7 +16,7 @@ export const programPageDataFetch = async (): Promise<ProgramsPageData> => {
     };
 };
 
-export const fetchProgramBySlug = async (slug: string): Promise<Program> => {
-    const response = await axiosInstance.get<Program>(`${API_ROUTES.PROGRAMS.BY_SLUG}/${slug}`);
+export const fetchProgramBySlug = async (slug: string): Promise<DetailedProgram> => {
+    const response = await axiosInstance.get<DetailedProgram>(`${API_ROUTES.PROGRAMS.BY_SLUG}/${slug}`);
     return response.data;
 };
