+++ src/types/common/programs.ts
@@ -1,0 +1,5 @@
+export enum ContentType {
+    Title = 0,
+    Description = 1,
+    Image = 2,
+}
